Check some business requirement

+++ app/config.go
@@ -3,12 +3,12 @@ package app
 import "github.com/spf13/viper"
 
 type Config struct {
-	MaxStringLength int64
+	MaxStringLength int
 }
 
 func InitConfig() (*Config, error) {
 	config := &Config{
-		MaxStringLength: viper.GetInt64("App.MaxStringLength"),
+		MaxStringLength: viper.GetInt("App.MaxStringLength"),
 	}
 	if config.MaxStringLength == 0 {
 		config.MaxStringLength = 10
